\name{templates}
\alias{template}
\alias{preamble}
\alias{configure}
\non_function{}

\title{ LaTex template files }

\description{
Makes available template files for the package \code{inst/extdata} directory.
}

\details{
Loads data.
}

\examples{
template()
preamble()
configure()
}

\keyword{ datasets }

