\name{voucher-package}
\alias{voucher-package}
\alias{voucher}
\docType{package}

\title{ Create herbarium labels for vegsoup objects }

\description{
Functions to typeset herbarium labels using LaTex from data extracted from a (\href{http://r-forge.r-project.org/projects/vegsoup/}{vegsoup}) object.
}

\details{
\tabular{ll}{
Package: \tab voucher\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2013-12-11\cr
License: \tab GPL (>= 2)\cr
}

A bunch of utility functions.
}

\author{
Maintainer:  Roland Kaiser <kardinal.eros@gemail.com>
}

\references{
Culley, T. M. (2013). Why vouchers matter in botanical research. Applications in Plant Sciences, 1:1300076.
}

\keyword{ package }
\seealso{
\code{\link[vegsoup:vegsoup-package]{vegsoup}}
}
\examples{
# Nothing here yet
}
