% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_vouch}
\alias{use_vouch}
\title{Initialize vouch in the current project}
\usage{
use_vouch()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Creates a starter \code{VOUCHED.td} file at \code{.github/VOUCHED.td} and ensures
\code{.github} is ignored in package builds by adding \verb{^\\\\.github$} to
\code{.Rbuildignore} when needed.
}
\details{
If a vouch file already exists (\code{VOUCHED.td} or \code{.github/VOUCHED.td}), this
function exits without making changes.
}
\examples{
\dontrun{
use_vouch()
}

}
